Split out only differing functions into PurchaseBC4 and PurchasesBC5 API
