+++ api-tester/src/main/java/com/revenuecat/apitester/java/PurchasesAPI.java
@@ -160,13 +160,4 @@ final class PurchasesAPI {
         final LogHandler handler = Purchases.getLogHandler();
     }
 
-    static void check(final BillingFeature feature) {
-        switch (feature) {
-            case SUBSCRIPTIONS:
-            case SUBSCRIPTIONS_UPDATE:
-            case PRICE_CHANGE_CONFIRMATION:
-            case IN_APP_ITEMS_ON_VR:
-            case SUBSCRIPTIONS_ON_VR:
-        }
-    }
 }
