+++ api-tester/src/main/java/com/revenuecat/apitester/kotlin/PurchasesAPI.kt
@@ -231,13 +231,4 @@ private class PurchasesAPI {
         val handler = Purchases.logHandler
     }
 
-    fun check(billingFeature: BillingFeature) {
-        when (billingFeature) {
-            BillingFeature.PRICE_CHANGE_CONFIRMATION,
-            BillingFeature.SUBSCRIPTIONS,
-            BillingFeature.SUBSCRIPTIONS_UPDATE,
-            BillingFeature.IN_APP_ITEMS_ON_VR,
-            BillingFeature.SUBSCRIPTIONS_ON_VR -> {}
-        }.exhaustive
-    }
 }
